--- api-server/src/main/java/com/lastone/apiserver/exception/MemberAlreadyExistException.java
@@ -1,13 +1,0 @@
-package com.lastone.apiserver.exception;
-
-import com.lastone.core.common.response.ErrorCode;
-
-public class MemberAlreadyExistException extends MyPageException {
-    public MemberAlreadyExistException(String message, ErrorCode errorCode) {
-        super(message, errorCode);
-    }
-
-    public MemberAlreadyExistException(ErrorCode errorCode) {
-        super(errorCode);
-    }
-}
