--- api-server/src/main/java/com/lastone/apiserver/exception/MemberNotFountException.java
@@ -1,13 +1,0 @@
-package com.lastone.apiserver.exception;
-
-import com.lastone.core.common.response.ErrorCode;
-
-public class MemberNotFountException extends MyPageException {
-    public MemberNotFountException(String message, ErrorCode errorCode) {
-        super(message, errorCode);
-    }
-
-    public MemberNotFountException(ErrorCode errorCode) {
-        super(errorCode);
-    }
-}
