+++ api-server/src/main/java/com/lastone/apiserver/exception/mypage/MemberNotFountException.java
@@ -1,0 +1,9 @@
+package com.lastone.apiserver.exception.mypage;
+
+import com.lastone.core.common.response.ErrorCode;
+
+public class MemberNotFountException extends MyPageException {
+    public MemberNotFountException() {
+        super(ErrorCode.MEMBER_NOT_FOUND);
+    }
+}
