+++ api-server/src/main/java/com/lastone/apiserver/exception/mypage/MyPageException.java
@@ -1,13 +1,9 @@
-package com.lastone.apiserver.exception;
+package com.lastone.apiserver.exception.mypage;
 
 import com.lastone.core.common.response.ErrorCode;
 import com.lastone.core.common.exception.LastOneException;
 
 public class MyPageException extends LastOneException {
-    public MyPageException(String message, ErrorCode errorCode) {
-        super(message, errorCode);
-    }
-
     public MyPageException(ErrorCode errorCode) {
         super(errorCode);
     }
